+++ source/shell/sys/linux/kernel/socket/unix/vsf_linux_socket_unix.c
@@ -76,7 +76,7 @@ struct vsf_linux_socket_unix_priv_t {
 /*============================ GLOBAL VARIABLES ==============================*/
 /*============================ PROTOTYPES ====================================*/
 
-static int __vsf_linux_socket_unix_fcntl(vsf_linux_fd_t *sfd, int cmd, long arg);
+static int __vsf_linux_socket_unix_fcntl(vsf_linux_fd_t *sfd, int cmd, uintptr_t arg);
 static ssize_t __vsf_linux_socket_unix_read(vsf_linux_fd_t *sfd, void *buf, size_t count);
 static ssize_t __vsf_linux_socket_unix_write(vsf_linux_fd_t *sfd, const void *buf, size_t count);
 static int __vsf_linux_socket_unix_close(vsf_linux_fd_t *sfd);
@@ -114,7 +114,7 @@ const vsf_linux_socket_op_t vsf_linux_socket_unix_op = {
 
 /*============================ IMPLEMENTATION ================================*/
 
-static int __vsf_linux_socket_unix_fcntl(vsf_linux_fd_t *sfd, int cmd, long arg)
+static int __vsf_linux_socket_unix_fcntl(vsf_linux_fd_t *sfd, int cmd, uintptr_t arg)
 {
     vsf_linux_socket_unix_priv_t *priv = (vsf_linux_socket_unix_priv_t *)sfd->priv;
     switch (cmd) {
