+++ source/shell/sys/linux/kernel/socket/vsf_linux_socket.c
@@ -191,7 +191,7 @@ const char * inet_ntop(int af, const void *src, char *dst, socklen_t size)
 
 // fcntl, dedicated driver can over-write this common version
 WEAK(__vsf_linux_socket_inet_fcntl)
-int __vsf_linux_socket_inet_fcntl(vsf_linux_fd_t *sfd, int cmd, long arg)
+int __vsf_linux_socket_inet_fcntl(vsf_linux_fd_t *sfd, int cmd, uintptr_t arg)
 {
     switch (cmd) {
     case F_SETFL:
